ft[utils]: Moves all utils to one folder in models.
